Move instantiate tests into test/instantiate/
